pyproject.toml
[project]
name = "engineering-decision-logger"
version = "0.2.0"
description = "A tool to log and manage engineering decisions using ADR format, Git integration, and a modern RS Digital GUI."
authors = [
    { name = "Robert Salinas" }
]
dependencies = [
    "customtkinter>=5.0.0",
    "typer[all]>=0.9.0",
    "sqlmodel>=0.0.14",
    "rich>=13.0.0",
    "gitpython>=3.1.0",
    "jinja2>=3.1.0",
    "python-slugify>=8.0.0",
    "fastapi>=0.100.0",
    "uvicorn>=0.20.0",
    "mkdocs-material>=9.0.0",
    "pyyaml>=6.0.0"
]
requires-python = ">=3.11"

[project.scripts]
edl = "src.cli:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
